Include data files and lsq models; update pathing
