+++ spectacle/data/cos.ecsv
@@ -1,0 +1,107 @@
+# %ECSV 0.9
+# ---
+# datatype:
+# - {name: index, datatype: int64, description: Index of kernel value}
+# - {name: value, datatype: float64, description: Kernel value}
+index value
+-50 8.95741269841e-05
+-49 9.71033730159e-05
+-48 0.000105633730159
+-47 0.000115231746032
+-46 0.00012603531746
+-45 0.000138233333333
+-44 0.000152091269841
+-43 0.000167882142857
+-42 0.000186013888889
+-41 0.000206763888889
+-40 0.00023046468254
+-39 0.000257469444444
+-38 0.000288477777778
+-37 0.000324491666667
+-36 0.000366572619048
+-35 0.000415809920635
+-34 0.000472916269841
+-33 0.00053863531746
+-32 0.000613946825397
+-31 0.000700473809524
+-30 0.000800205952381
+-29 0.000914940079365
+-28 0.00104543015873
+-27 0.00119157222222
+-26 0.00135255555556
+-25 0.00152807539683
+-24 0.00171618253968
+-23 0.00191355952381
+-22 0.0021161031746
+-21 0.00232332539683
+-20 0.0025414047619
+-19 0.00278194444444
+-18 0.00305919444444
+-17 0.00338529761905
+-16 0.00376579365079
+-15 0.00419898412698
+-14 0.00468388888889
+-13 0.00523658730159
+-12 0.00590370634921
+-11 0.00676173809524
+-10 0.00791490079365
+-9 0.00951524603175
+-8 0.0118233888889
+-7 0.0153451587302
+-6 0.0210300396825
+-5 0.0303469047619
+-4 0.0447793650794
+-3 0.0643869047619
+-2 0.0860518253968
+-1 0.103360714286
+0 0.109557460317
+1 0.102020079365
+2 0.0840294047619
+3 0.0622851984127
+4 0.0427194444444
+5 0.0282088492063
+6 0.0188404365079
+7 0.0133118730159
+8 0.0101249603175
+9 0.00819181349206
+10 0.00690545238095
+11 0.00597822222222
+12 0.00526676984127
+13 0.00468511904762
+14 0.00418128174603
+15 0.00372967063492
+16 0.0033226984127
+17 0.00295932142857
+18 0.00263878968254
+19 0.00235970238095
+20 0.00211854365079
+21 0.00190951190476
+22 0.00172723015873
+23 0.00156798809524
+24 0.001429
+25 0.00130673373016
+26 0.00119608690476
+27 0.00109135793651
+28 0.000989727380952
+29 0.000890991666667
+30 0.00079681984127
+31 0.000709594047619
+32 0.000631046825397
+33 0.000561691269841
+34 0.000501157142857
+35 0.000448762698413
+36 0.000403791269841
+37 0.000365252777778
+38 0.000332014285714
+39 0.000302819047619
+40 0.000276757539683
+41 0.0002533
+42 0.000231972222222
+43 0.000212429365079
+44 0.000194296428571
+45 0.000177523015873
+46 0.000162054365079
+47 0.000148095634921
+48 0.000135631349206
+49 0.000124665079365
+50 0.000114981349206
